last

+++ app/Http/Controllers/BankAccount/BankAccountController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers\BankAccount;
 
 use App\Http\Requests\BankAccount\BankAccountRequest;
-use App\Http\Requests\BankAccount\UpdateBankAccountRequest;
 use App\Http\Resources\BankAccount\BankAccountResource;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
@@ -17,15 +16,7 @@ class BankAccountController extends Controller
 
     public function find($id)
     {
-        // $model = cacheGet('countries_' . $id);
-        // if (!$model) {
-        //     $model = $this->modelInterface->find($id);
-        //     if (!$model) {
-        //         return responseJson(404, __('message.data not found'));
-        //     } else {
-        //         cachePut('countries_' . $id, $model);
-        //     }
-        // }
+
         $model = $this->modelInterface->find($id);
         if (!$model) {
             return responseJson(404, __('message.data not found'));
@@ -35,15 +26,7 @@ class BankAccountController extends Controller
 
     public function all(Request $request)
     {
-        // if (count($_GET) == 0) {
-        //     $models = cacheGet('countries');
-        //     if (!$models) {
-        //         $models = $this->modelInterface->all($request);
-        //         cachePut('countries', $models);
-        //     }
-        // } else {
-        //     $models = $this->modelInterface->all($request);
-        // }
+
         $models = $this->modelInterface->all($request);
         return responseJson(200, 'success', BankAccountResource::collection($models['data']), $models['paginate'] ? getPaginates($models['data']) : null);
     }
@@ -77,21 +60,110 @@ class BankAccountController extends Controller
         return responseJson(200, 'success', \App\Http\Resources\Log\LogResource::collection($logs));
     }
 
+    // public function delete($id)
+    // {
+    //     $model = $this->modelInterface->find($id);
+    //     if (!$model) {
+    //         return responseJson(404, __('message.data not found'));
+    //     }
+
+    //     if ($model->hasChildren()) {
+    //         return responseJson(400, __("this item has children and can't be deleted remove it's children first"));
+    //     }
+
+    //     $this->modelInterface->delete($id);
+
+    //     return responseJson(200, 'success');
+    // }
+
+    // public function bulkDelete(Request $request)
+    // {
+    //     foreach ($request->ids as $id) {
+    //         $model = $this->modelInterface->find($id);
+    //         $arr = [];
+    //         if ($model->hasChildren()) {
+    //             $arr[] = $id;
+    //             continue;
+    //         }
+    //         $this->modelInterface->delete($id);
+    //     }
+    //     if (count($arr) > 0) {
+    //         return responseJson(400, __('some items has relation can\'t delete'));
+    //     }
+    //     return responseJson(200, __('Done'));
+    // }
+
     public function delete($id)
     {
         $model = $this->modelInterface->find($id);
         if (!$model) {
             return responseJson(404, __('message.data not found'));
         }
+
+        $relationsWithChildren = $model->hasChildren();
+
+        if (!empty($relationsWithChildren)) {
+            $errorMessages = [];
+            foreach ($relationsWithChildren as $relation) {
+                $relationName = $this->getRelationDisplayName($relation['relation']);
+                $childCount = $relation['count'];
+                $childIds = implode(', ', $relation['ids']);
+                $errorMessages[] = "This item has {$childCount} {$relationName} (IDs: {$childIds}) and can't be deleted. Remove its {$relationName} first.";
+            }
+            return responseJson(400, $errorMessages);
+        }
+
         $this->modelInterface->delete($id);
+
         return responseJson(200, 'success');
     }
 
+
     public function bulkDelete(Request $request)
     {
+        $itemsWithRelations = [];
+
         foreach ($request->ids as $id) {
+            $model = $this->modelInterface->find($id);
+
+            $relationsWithChildren = $model->hasChildren();
+            if (!empty($relationsWithChildren)) {
+                $itemsWithRelations[] = [
+                    'id' => $id,
+                    'relations' => $relationsWithChildren,
+                ];
+                continue;
+            }
+
             $this->modelInterface->delete($id);
         }
+
+        if (count($itemsWithRelations) > 0) {
+            $errorMessages = [];
+            foreach ($itemsWithRelations as $item) {
+                $itemId = $item['id'];
+                $relations = $item['relations'];
+
+                $relationErrorMessages = [];
+                foreach ($relations as $relation) {
+                    $relationName = $this->getRelationDisplayName($relation['relation']);
+                    $childCount = $relation['count'];
+                    $childIds = implode(', ', $relation['ids']);
+                    $relationErrorMessages[] = "Item with ID {$itemId} has {$childCount} {$relationName} (IDs: {$childIds}) and can't be deleted. Remove its {$relationName} first.";
+                }
+
+                $errorMessages[] = implode(' ', $relationErrorMessages);
+            }
+
+            return responseJson(400, $errorMessages);
+        }
+
         return responseJson(200, __('Done'));
+    }
+
+    private function getRelationDisplayName($relation)
+    {
+        $displayableName = str_replace('_', ' ', $relation);
+        return ucwords($displayableName);
     }
 }
