+++ app/Http/Controllers/Unit/UnitController.php
@@ -100,17 +100,91 @@ class UnitController extends Controller
             return responseJson(404, __('message.data not found'));
         }
 
+        $relationsWithChildren = $model->hasChildren();
+
+        if (!empty($relationsWithChildren)) {
+            $errorMessages = [];
+            foreach ($relationsWithChildren as $relation) {
+                $relationName = $this->getRelationDisplayName($relation['relation']);
+                $childCount = $relation['count'];
+                $childIds = implode(', ', $relation['ids']);
+                $errorMessages[] = "This item has {$childCount} {$relationName} (IDs: {$childIds}) and can't be deleted. Remove its {$relationName} first.";
+            }
+            return responseJson(400, $errorMessages);
+        }
+
         $this->modelInterface->delete($id);
 
         return responseJson(200, 'success');
     }
 
+
     public function bulkDelete(Request $request)
     {
+        $itemsWithRelations = [];
+
         foreach ($request->ids as $id) {
+            $model = $this->modelInterface->find($id);
+
+            $relationsWithChildren = $model->hasChildren();
+            if (!empty($relationsWithChildren)) {
+                $itemsWithRelations[] = [
+                    'id' => $id,
+                    'relations' => $relationsWithChildren,
+                ];
+                continue;
+            }
+
             $this->modelInterface->delete($id);
         }
+
+        if (count($itemsWithRelations) > 0) {
+            $errorMessages = [];
+            foreach ($itemsWithRelations as $item) {
+                $itemId = $item['id'];
+                $relations = $item['relations'];
+
+                $relationErrorMessages = [];
+                foreach ($relations as $relation) {
+                    $relationName = $this->getRelationDisplayName($relation['relation']);
+                    $childCount = $relation['count'];
+                    $childIds = implode(', ', $relation['ids']);
+                    $relationErrorMessages[] = "Item with ID {$itemId} has {$childCount} {$relationName} (IDs: {$childIds}) and can't be deleted. Remove its {$relationName} first.";
+                }
+
+                $errorMessages[] = implode(' ', $relationErrorMessages);
+            }
+
+            return responseJson(400, $errorMessages);
+        }
+
         return responseJson(200, __('Done'));
     }
+
+    private function getRelationDisplayName($relation)
+    {
+        $displayableName = str_replace('_', ' ', $relation);
+        return ucwords($displayableName);
+    }
+
+    // public function delete($id)
+    // {
+    //     $model = $this->modelInterface->find($id);
+    //     if (!$model) {
+    //         return responseJson(404, __('message.data not found'));
+    //     }
+
+    //     $this->modelInterface->delete($id);
+
+    //     return responseJson(200, 'success');
+    // }
+
+    // public function bulkDelete(Request $request)
+    // {
+    //     foreach ($request->ids as $id) {
+    //         $this->modelInterface->delete($id);
+    //     }
+    //     return responseJson(200, __('Done'));
+    // }
 
 }
